Do not set the divider pos during adjust ime and minimized interaction
When launching an app in splitscreen with while ime is up will trigger
an animation that is trying to minimize the docked stack (this bug is
tracked in b/36603383) with the end result in incorrect bounds for the
docked and fullscreen tasks (leaving visual glitches in their apps).
This was caused by setting the divider position while there was
interaction with the minimized state. Therefore ignore setting divider
while adjusting for ime and minimized interaction for the tasks to end
with the correct bounds.

Fixes: 38428386
Test: manual - dock something in split screen, use the launcher search
to bring up ime, launch any app (like gmail)
Change-Id: Ic3547b02a39f40698db07d1c0946e0991b69f5d3

Former-commit-id: f3589c7a2231194094af1572e91ded7b1342fce0