Reset mHiddenSetFromTransferredStartingWindow when app set visibility
ag/2276103 fixed an issue where an app token was left in the visible state
due to it getting a tranferred starting window that is later transferred out.
However, we were not resetting the state of the variable we use to track this
when the app actually sets a visibility state which can cause the app to
remain in an hidden state.

Fixes: 62011969
Bug: 38317227
Test: Cold launch Docs apps and change the orientation and make sure the app
is still displayed.
Test: Launch Audible app music player and verify the LibraryActivity
token isn't in the visible state.

Change-Id: I9ff434efe8d29b32021b9881bf36e69de700e70e

Former-commit-id: 9e4721f53029f2d3386c11a46dc8b3289964abf3