WifiManager: add iface name and mode labels
Connectivity service needs to get the interface name and intended ip
configuration mode for softap.
This CL adds the interface name and softap target mode variables to
use to insert the extra info into the wifi ap state change broadcast.

Note: additional CLs are needed to verify the iface name and mode.  The
mode tracking is still in progress.

Bug: 62076211
Test: frameworks/base/wifi/tests/runtests.sh
Test: frameworks/opt/net/wifi/tests/wifitests/runtests.sh
Test: manually verified the name in the broadcast
Change-Id: I64e1d13a9e422945fe9c4c268c206da365d0a3eb

Former-commit-id: 4ac2cab8fd8cb6ef877f4e5c37f554f1a09b79dc