Fixed an issue where the panel could be stuck tracking
Multiple issues are fixed that could lead to the panel being
stuck on the lockscreen. The easiest way to reproduce it was
to swipe up while on ambient display, which could easily happen
in a pocket.
This also adds some safeguards, such that it won't happen again
and ensures that notifications can't be swiped in AOD.

Test: add notifications in AOD, try swiping on them
Change-Id: I8ba0ebe72c3a2734b59443f3b93dbe5f1837cbbd
Fixes: 38486627

Former-commit-id: be2c443e21410bb27decec5d2aea60ef37f10c05