Handle not having main window in createThumbnailAppAnimator.
A recent code change associated the SurfaceControl with the main
window owner id. However, it is possible for the an AppWindowToken to
not have a main window. This case is handled later in the method.

This CL restores the original behavior of using the calling uid in
the cases no main window is present.

Change-Id: Iad69fe383c2208c1db523c8b4601a8f927f9318a
Fixes: 62096254
Test: go/wm-smoke

Former-commit-id: 6db4d15d726b8830ad56766352a79fe0417cb9c2