Handle not having main window in createThumbnailAppAnimator.
A recent code change associated the SurfaceControl with the main
window owner id. However, it is possible for the an AppWindowToken to
not have a main window. This case is handled later in the method.

This CL restores the original behavior of using the calling uid in
the cases no main window is present.

Change-Id: I8255be9e0d68adc75fda0947c64f869b7eeb76c9
Fixes: 62096254
Test: go/wm-smoke

Former-commit-id: e6c25d4a6f35449a8420f0d1e87529426786ad60