Handle renamed packages during default grants
If a device has a privileged app on the system image which was
updated and a future OTA renames the app package (supported only for
privileged apps) we end up with a disabled system package setting
with no package information. Since we are not doing a null check
of the package we get from the disabled package settings during
default grants we crash leaving the system in a bad state.

Test: Manual
Change-Id: Id7255c92a4a52e368266739cd331977256eee2d3
Fixes: 62376260

Former-commit-id: 7e79d354a1f2426ede5bd08384c0dbda665e0cac