Replace mockito with a mock class in DynamicLayoutTest
Previously, a mocked object was used incorrectly in the test. It was
attached twice as a span, which would result in the second identical
span causing the first span to get removed.

Now we use two different instances of a mock class, to make sure we
are attaching different spans.

Fixes: 62217995
Test: adb shell am instrument -w -e package android.text com.android.frameworks.coretests/android.support.test.runner.AndroidJUnitRunner

Change-Id: Ib9f7a467410d28280d77a1023e0fc328c652e2af

Former-commit-id: de0353dbc70e4e95d2892609cbd26904fbacd20a