Replace fonts of unknown origin
There are fonts that we are not sure where they came from,
replace with fonts we created ourselves.

The res/font/ folder does not build with gradle when there is
a ttx file in it, so fonts there have been copied from elsewhere
in the tree and contain copyright and licensing inside the font.
There is also a new readme in the root folder for the test projects
using these to point this out.

Bug: 37726227
Test: bit FrameworksCoreTests:android.content.res.FontResourcesParserTest
Change-Id: I43364abb9d466e9673902246b0d7c3043af619c1

Former-commit-id: d2e7a012394441af21c5a8b9f1ff75816c56dad8