Prioritize reused activity's stack over computed stack.
It is possible that an activity in one stack may be reused in
another. For example, if an activity is started from a launcher
intent, but then is started from a home intent (from a
ResolverActivity). We currently do not move the activity, leading
to a inconsistency as the window manager proceeds to position the
task in the focus stack.

This changelist addresses the issue by using the reused activity's
stack rather than the computed stack.

Change-Id: Ie8a099e57e05a20b247bd0c97df8cda69e17c1bb
Fixes: 62402289
Test: go/wm-smoke

Former-commit-id: a19b5ad748ec8d495ee958e7ac48b9e6cba8e5b4