LocalLog: Use LocalDateTime.now() instead of a Calendar.
Calendars are cloned once per format argument they're used for and cloning
a calendar is quite expensive. The line modified by this change used to
create dozens of objects, most of them pointlessly cloned.

Instead, we LocalDateTime, which is immutable and will be used directly.
Moreover, we can rely on its toString method which is specified to
format dates the same way we were doing in this class.

Test: manual
Change-Id: Ia2f6f5eb4f79252663c22b5a81d747c928771ee9

Former-commit-id: 6229d990828aae67089c642fb47ac5b6daaacd48