Eagerly serialize historical sessions
mHistoricalSessions maintained a strong reference to a PackageInstallerSession,
which in turn kept references to Bitmaps and other heavy-weight objects around.

Since this field is primarily used for debugging, this change replaces it with
a String dump of the session in question. Each dump takes about 600bytes, which
is comparable to the sizes of the un-serialized raw objects.

Bug: 62485552
Test: Manual

Change-Id: I4949a64b538ab4a97384f4f8bc9a6ef155a4b128

Former-commit-id: a22a766ea3eb14466221ac5a29ac7c298446ca5e