Fix issue #62390590:  SecurityException in JobIntentService$...
...JobServiceEngineImpl$WrapperWorkItem.complete

When a job is in the process of stopping a job, we should no
longer allow new work to be dispatched for it.

Also there was an issue with how we determine wether a caller
is valid for the current job -- this was only based on the uid
of the currently running job, but of course that context could
be re-used for a new job of the same uid.  Instead, we now create
a different callback binder for each client, so we can identify
the exact client each time it calls back.  (This also allows us
to hang information about why the job last stopped on that
client state, so we can always report it.)

Finally make a bunch of classes final that should have been.

Test: bit CtsJobSchedulerTestCases:*

Change-Id: I1b00dd2da710414dd2898c4d39a5c528d54b95ea

Former-commit-id: 6466c1cc5efc4ff05fabdd670cf78a6a7ca45afb