Define a meta-data attribute for a carrier icon.
This is just a centralized key that a carrier app and various eSIM
infrastructure can agree upon. It may be used to allow the carrier app
to show a custom icon instead of its launcher icon when some UI is
rendering the identity of the carrier. Use of this meta-data is
entirely up to callers.

Fixes: 62580015
Test: TreeHugger
Change-Id: Iaaab8cb8b5108d92b8f687d213daa7917f53fe65

Former-commit-id: f0d8c896c7f0d40bb36a05a87fc3c56bfeddc9f1