Don't even think about changing keyguard transit
When camera was launched with a lockscreen wallpaper set, the
wallpaper target was launcher in that case, which was also in
mClosingApps because it was first getting shown by keyguard
exit but then immediately hidden by starting the camera,
before the transition started.

Now since lockscreen wasn't the wallpaper target, launcher was
already for some reason, and we changed the transit to
WALLPAPER_CLOSE as a window with the wallpaper target was in
mClosingApps.

Fix this by never ever changing away from keyguard transits.

Test: go/wm-smoke
Test: ActivityManagerTransitionSelectionTests
Test: Set lockscreen wallpaper, set animation duration scale to
0.5, insert a random sleep statement in SystemUI, launch
camera from screen off while in trusted state and camera
wasn't running before.
Fixes: 37677242
Change-Id: I984b66d7f117034f3d55591284dd822b5ec76cbd

Former-commit-id: a69243a5475c3cecc410c3328f221bab220cab8a