Fix searching for overridden methods.
If the superclass is an abstract class, we want to search it for
possibly overridden methods; we don't want to search ourselves.

This bug resulted in (incorrectly) thinking that overridden "default"
methods from interfaces were standalone methods on the class.  (The
expected behavior has always been to omit any overridden methods from
the public API surface area to avoid redundant noise.)

Test: make update-api
Bug: 62675475
Change-Id: Ic92544908692bd9af329d302165e18c7b4738e43

Former-commit-id: 257529dcae079055d51cb8eb42b249512ae41f36