Do not restore same autofill id twice
View's are restored based on their id. The autofill id is part of the
id's state. Unfortunately the ids of the view are not neccessary
unique. In this case two views would store state of for the same id.
Also two views would be restored frome the same state. This causes
duplicate autofill ids which is not allowed. Hence make sure that for a
single state, the autofill id is only restored once.

Change-Id: I8d950486cb21d9f6afe6d0f5f668799305ae73e7
Fixes: 62658714
Test: cts-tradefed run cts-dev -m CtsAutoFillServiceTestCases

Former-commit-id: cb797813565792431775b69cec98a77ddc9added