Cleanup a.t.InstrumentationTestRunner
ARGUMENT_TEST_CLASS, ARGUMENT_TEST_PACKAGE, ARGUMENT_TEST_SIZE_PREDICATE
and ARGUMENT_DELAY_MSEC were made package private because they are not
part of the API and are only used from within the same package.

The JavaDoc comment was changed to a multi-line comment because it is
not and should not be associated with any member.

Fixed a couple of minor JavaDoc issues.

Bug: 30188076
Test: make checkbuild
Change-Id: If28460aec2fd0080e0f5346a83abae393b7a9fb2

Former-commit-id: 575f9da647c2006d3e895adb2c418b51e50d3d63