Fixes getting the wrong app transition going home after minimized (1/2)
This is a side affect from ag/2125707. When minimizing and unminimizing
back to home would cause the same affect in the bug, where the app
transition is from activity to activity and not app. The issue was a
"updateAllDrawn" is normally called at the end of the surface
transaction and therefore removing this call (that would run inbetween
the transaction) should fix prematurely calculating the app transition
before setting the wallpaper target.

Fixes: 62727547
Test: go/wm-smoke
Test: cts/hostsidetests/services/activityandwindowmanager/util/run-test
CtsServicesHostTestCases
android.server.cts.ActivityManagerDockedStackTests#
testMinimizeAndUnminimizeThenGoingHome
Change-Id: Ia658cd7111ede386579eb5cd38c88c9e4a1bd312

Former-commit-id: 5d23afa9e55d526cbf86042d8efbb69250f25038