Use configuration delta to determine if still in orientation change.
We currently track this change in the WindowState with a variable
called mOrientationChanging. This value is set and cleared in a
variety of classes. It is possible for this value to be cleared
before processing an orientation changing frame. This can lead to us
prematurely unfreezing.

Since the orientation is present in the configuration, we can use the
orientation delta between the last and current frame's configuration
to determine whether we are in the middle of an orientation change.
The existing signal has been moved behind a setter/getter where the
latter is combined with the configuration state.

Bug: 62846907
Test: go/wm-smoke
Test: turn launcher rotation on, open dialer in portrait, turn off
      screen, rotate to landscape, unlock with fingerprint, press
      home while rotating. make sure layout is correct.

Change-Id: Ie0b78b76565125fd1cb036545cfb59f2a9075328

Former-commit-id: 2a0284b1a69b57cc239357c9a5fbb514cf2fc8b2