Ensure all WindowStates are considered before determining all drawn.
The drawn state of an AppWindowToken is calculated on demand in
updateAllDrawn. updateAllDrawn assumes that updateDrawnWindowStates
has been called on every child window beforehand. If a window has
been added but not had updateDrawnWindowStates called for it, the
results of updateAllDrawn will be wrong. This leads to situations
where we will determine we are ready to execute a transition to
home, yet not have a wallpaper activity.

This changelist addresses this issue by recording which WindowStates
have had updateDrawnWindowStates called on them and making sure all
present children have been processed in updateAllDrawn.

Bug: 62846907
Test: go/wm-smoke
Test: turn launcher rotation on, open dialer in portrait, turn off
      screen, rotate to landscape, unlock with fingerprint, press
      home while rotating. make sure transition is correct.

Change-Id: I3d129e686326d868862ae988725fb091a2c86e4e

Former-commit-id: d390deb8f564b9aeab0257dfefd8a051d8a57f7f