Remove a redundant synchronized block
This CL removes a redundant synchronized block, which was accidentally
added in the previous CL [1] that simplified
TextServicesManagerService (TSMS) by using RemoteCallbackList<E>.

Since SpellCheckerBindGroup#removeListener() internally acquires the
lock on TSMS#mSpellCheckerMap, there is no need to acquire the same
lock in the caller.

This CL is just a mechanical clean-up. No behavior change is intended.

 [1]: I38942765ed6bec6713757b1d5f325e7a633c2ba7
      8bee6c4d84eb5ce087abaef43001f514d871f75d [formerly f982e75a262db16e60fd3565638d399315dba761]

Bug: 35102403
Test: compile
Change-Id: I6d47e8cb6e301286bcb26922b74ee7cac2c5c1c5

Former-commit-id: f14fe14a3bd09ac60d70e33248eae88c4da08a66