Fix TimePicker RTL issues
The TimePicker (seen when creating an alarm in the Clock app) has
two bugs in landscape mode:
- The hours/minutes switch places (hours should always be to the left
of minutes, regardless of RTL)
- The layout is completely messed up in 12-hour mode (the radial time picker
does not appear on the screen)

This fix addresses both issues by setting the layoutDirection="ltr"
attribute to force a couple of the layouts to be unaffected in RTL mode.
Specifically, the hours/minutes layout and the overall layout will always be
presented left-to-right.

Bug: 62846484  TimePicker layout is broken in RTL landscape mode
Test: manual (tested with/without fix in normal, forded-RTL mode, and
with Hebrew as the first locale).

Change-Id: I8ef3f2706660f60c162ba3d324a18b8f370f67f8

Former-commit-id: e2dfa0e343b5a0babb050b08355e781da3bbd516