Update android.test.runner API
Bug 62675475 changed DocLava to add some missing methods to the API
definition file. This updates the API definition for the
android.test.runner API.

The reason this was not found when it was uploaded is because there is a
discrepancy between the checkapi target and the update-api. The checkapi
target will not detect that the update-api needs to make changes. This
issue has been reported.

Bug: 30188076
Test: make update-api && make checkapi
Change-Id: I598774d3239d58eb5c58ff78e24e8ca84078e64a

Former-commit-id: 09c29c1324b96e5404400947da7eeb9f0976353a