Fix RTL mobile signal drawing out of bounds in QS Edit
For some reason, the canvas is twice the size of the drawable's bounds,
and our RTL code was taking the whole canvas, shifting by its own width
and reversing the x-coords. When the canvas is too wide, it causes the
signal to draw outside of the bounds. To fix it, just shift the
coordinates by the width of the drawable itself. I.e., reverse in the
drawable's own bounds.

Note that there is still an issue where sometimes a non-reversed signal
drawable is used. This doesn't fix that.

Change-Id: I6cfda0ab6b9f50c969410dd4f89499b66854ebbe
Fixes: 63164522
Test: visual

Former-commit-id: ae5b7002ea206302abf106696d60f90ef2cf3e64