Fix the VR virtual display height
The display of the VR virtual display was set incorrectly in ag/2460110.
This was a part of debugging code mistakenly left behind, but it is
causing CTS tests to fail and also changes the resolution of virtual
display on standalone devices.

Bug: 62842592
Bug: 63339799
Test: run cts -m CtsServicesHostTestCases -t
android.server.cts.ActivityManagerDisplayTests#testVrActivityLaunch
passes with the change and restores original resolution on standalone
devices.

Change-Id: I6e77f9b3b09f02e411bab12bb69d8141312e5b64
Signed-off-by: Karthik Ravi Shankar <karthikrs@google.com>

Former-commit-id: 6d1575fc24986d1ee13f682e75175e4540ebe999