Use configuration delta to determine if still in orientation change.
We currently track this change in the WindowState with a variable
called mOrientationChanging. This value is set and cleared in a
variety of classes. It is possible for this value to be cleared
before processing an orientation changing frame. This can lead to us
prematurely unfreezing.

Since the orientation is present in the configuration, we can use the
orientation delta between the last and current frame's configuration
to determine whether we are in the middle of an orientation change.
The existing signal has been moved behind a setter/getter where the
latter is combined with the configuration state.

This reverts commit 3e1e285840a9af6b5e18df5b05a08557a2683bdd [formerly 31386ab88b48fd5475efc58155549ad4c7da7ad2]. It
addresses issues raised in b/63380251 where the orientation delta was
being considered for WindowStates that were not visible.

Bug: 62846907
Test: go/wm-smoke
Change-Id: I574efb8f9a458cb295f2ef14e10dbb91e8875bf4

Former-commit-id: 8c3cf38592e0e197606cf0ce1b1d6e85b7e55711