fix cellular captive portal reevaluation url
we apply redirection link as the url everywhere for cellular captive portal.
this url should only be used when load the portal page at the first place (when users
click on the notification) and shouldn't be used as the re-evaluation url
as it wont return the expected response code.
instead we should use special URL (generate_204) from connectivity
service which will return 204 if redirection is gone.

Test: manual test with cold and hot TMO SIM
Bug: 63351490
Bug: 62532802
Change-Id: I56bb03a58447c5c46758e692ff357d540ea4cd7b

Former-commit-id: c457a04d937b0f3b81fb49b78f4cde46059e1ac9