Fix splits class loader context for non dependant splits
If the app doesn't request for splits to be loaded in isolation or does
not declare inter-split dependencies, then all the splits are loaded in
the base apk class loader (in the order of they are defined).

Fix the class loader context passed to dex2oat to reflect the runtime
loading logic.

Bug: 38138251
Test: runtest -x
services/tests/servicestests/src/com/android/server/pm/dex/DexoptUtilsTest.java

Change-Id: Ia0623d38883ae244fd16c0afb053fef016bf260a

Former-commit-id: 305aeea38f96f0b94ad4be5cb979dd9cf98cf7df