AOD: Only force doze brightness when actually unlocking
Previously, we forced the doze brightness even if we acquired a bad fingerprint;
because the state got never reset, the keyguard can be stuck in a dark mode.

We do not need to force the brightness until we wake up though, so this can be
delayed until we know if the unlock is going to succeed.

Change-Id: I72f4a577134223831b1bb0baefdbc227d8e8bab4
Fixes: 63931486
Test: Set up fingerprint; go to AOD; tap wrong finger; press power button; verify screen is fully on and not dim.

Former-commit-id: b880afe7839388e9adc387ae6dd0605b7ec93583