Refactor TextView TextAppearance attributes reading
Attributes that can appear in a TextAppearance are read in 3
different places inside TextView:
- Within the TextAppearance in the constructor
- Within the Style in the constructor
- Within setTextAppearance

This change simplifies the attribute reading code so it only
happens in one place. Note the way TextView reads these attributes
is first populating them from the TextAppearance and then allowing
them to be overriden by the style. The same is accomplished here.

For testing purposes and for API consistency, this change also adds
the missing APIs
- isAllCaps
- isElegantTextHeight

Bug: 64053645
Test: bit CtsWidgetTestCases:.TextViewTest
Change-Id: Iad6c1c6515992a51020c656edabea4e2927572f3

Former-commit-id: d2781160f1c5900565a29043da1e5848c6b98bda