Ensure that all displays are updated upon any changes to rotation.
There's generally only ever 1 display on most devices, so this code
has no change. If any additional displays have rotation support, their
displays will be updated accordingly with this change. At the moment,
this only affect VR's virtual display.

This fixes a bug where the Activity on a virtual display never ran
through PhoneWindowManager's rotation policy until it came across an
Activity that had any value for screenOrientation.
The result would be that the activity would display the natural
landscape orientation of the virtual display instead of the portrait
preference that is written into PhoneWindowManager. This change ensures
that the rotation is confirmed with PhoneWindowManager during Activity
layout.

Test: Before change - Run virtual display and see that it doesn't get
it's orientation set via PhoneWindowManager's policy.  After change -
see that the orientation is updated accordingly as part of this change.
Bug: 63289312, 62862109

Change-Id: I7de7aea790fd11daba36097227e13874bd9aacaa

Former-commit-id: 34b0991640ad2263f493f09df991dc54d7fee579