Dismiss recents after switching users
Switching users was causing Recents to load the other user's tasks. This
is bypassed by dismissing Recents when user is changed to match to how
turning the screen off dismisses Recents. When a user is switched, home
is shown after keyguard is unlocked. Then when Recents is launched again
it will load the usual tasks by the user.

Fixes: 63336488
Test: make 2nd user, open recents with at least one task, switch to 2nd
user and switch back, unlock keyguard to recents

Change-Id: I46dcec9954007068ff28cd0671a4242aa009bd2a

Former-commit-id: 3817fd201199d369405588656bb8d8cd679e1aa0