Remove Conferenceable update squelching
In an effort to remove redundant updates to Telecom,
conferenceable connection updates for a conference
were being removed if the list of connections hasn't
changed.

For GSM, when a new outgoing call was placed, it was
added to the conference controller before being added
to the list of connections. The conference controller
was updating the conferenceable connections for that
conference with a connection that wasn't being tracked
yet, so it was not notifying telecom of the connection.
It was still added to the list of conferenceable
connections in the conference, however, so later
conferenceable updates were being disregarded before
being sent to Telecom.

This change removes conferenceable update squelching
so that we always update telecom with the correct
conferencable connections for a conference.

Bug: 37989809
Test: Manual, Procedure in Bug, unit tests
Change-Id: I62960df9db27f0b303431aef99e73ef87498d374

Former-commit-id: 8c6f0cab335df7463b2a734ce63065e7c9d72a16