Use normal API for legacy wallpaper restore
No longer do we play crazy rename-into-place tricks!  Just take
the restored image and treat it as an ordinary new system wallpaper,
and let the system machinery do all its proper work.  While we're at
it, we get rid of the now-spurious shenanigans about backing up the
wallpaper under the 'android' package, an overdue and welcome bit
of thinning.

In addition, we now fully migrate a legacy wallpaper image to have
both a source and a display crop in place.  We were previously
not generating the equivalent source image, which made N+ backups
unable to store the image properly.

Fix 64269208
Fix 62809439
Test: manual (set wallpaper under M, backup, flash to ToT,
      force restore)

Change-Id: If9b26b777f0fda95cf37c3c790a3fa4e82ff0340

Former-commit-id: 9f22443b4e3ac1ca1763a25a186177809ec1c3cb