AOD: Unregister touchscreen sensors while pulsing
While we're pulsing, the touch screen is fully on. Keeping touchscreen
sensors registered can lead to the sensorhub and kernel touch driver
racing on who gets to control the touchscreen; the losing party may
also logspam and consume increased power.

To avoid this, we unregister touchscreen sensors when turning the screen
on.

Change-Id: Ie236d0a40c92d7fc544bd34ff883e613915c8e46
Fixes: 64160164
Test: receive notification; adb shell dumpsys sensorservice, verify double tap sensor is not registered; verify it is again registered after swiping away the notification.

Former-commit-id: 98d3198b5ced6b371067d0ff6166aaffff1dd512